VFB-315 organise file system, refactor duplicate calculations, make couplings in logic more explicit
