+++ src/app/parcels/batch/EmptyData.ts
@@ -1,4 +1,4 @@
-import { BatchEditData, OverrideData } from "@/app/parcels/batch/BatchTypes";
+import { BatchEditData, OverrideData } from "@/app/parcels/batch/batchTypes";
 
 export const emptyBatchEditData: BatchEditData = {
     client: {
