+++ src/app/parcels/batch/FieldValidationFunctions.ts
@@ -1,8 +1,8 @@
 import { GridEditCellProps, GridPreProcessEditCellProps } from "@mui/x-data-grid";
-import { phoneNumberRegex } from "@/common/format";
+import { phoneNumberRegex, phoneNumberFormatSymbolsRegex } from "@/common/format";
 
 export const phoneNumberValidation = (params: GridPreProcessEditCellProps): GridEditCellProps => {
-    const unformattedInput = params.props.value.replaceAll(/[\s-()]/g, "");
+    const unformattedInput = params.props.value.replaceAll(phoneNumberFormatSymbolsRegex, "");
     const hasError = unformattedInput.match(phoneNumberRegex) === null;
     return {
         ...params.props,
